hv: fix type conversion violations
This patch fixes the following violations:
 1. Implicit conversion: actual to formal param
 2. Value is not of appropriate type
 3. No cast for widening complex int expression
 4. Widening cast on complex integer expression
 5. Narrower int conversion without cast.

Tracked-On: #861

+++ hypervisor/arch/x86/guest/vlapic.c
@@ -317,7 +317,7 @@ set_expiration(struct acrn_vlapic *vlapic)
 	if ((tmicr == 0U) || (divisor_shift > 8U)) {
 		ret = false;
 	} else {
-		delta = tmicr << divisor_shift;
+		delta = (uint64_t)tmicr << divisor_shift;
 		timer = &vtimer->timer;
 
 		if (vlapic_lvtt_period(vlapic)) {
@@ -1163,7 +1163,7 @@ vlapic_get_cr8(const struct acrn_vlapic *vlapic)
 	uint32_t tpr;
 
 	tpr = vlapic_get_tpr(vlapic);
-	return (uint64_t)(tpr >> 4U);
+	return ((uint64_t)tpr >> 4UL);
 }
 
 static void
